requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pytest
pytest-mock
httpx
cryptography
email-validator
python-jose[cryptography]
